omap: iommu: generic iommu api migration
Migrate OMAP's iommu to the generic iommu api, so users can stay
generic, and non-omap-specific code can be removed and eventually
consolidated into a generic framework.

Tested on both OMAP3 and OMAP4.

Note: patch set is not bisectable, and is splitted at this point merely
to ease the review. later iterations will squash relevant patches to fix
this.

Signed-off-by: Ohad Ben-Cohen <ohad@wizery.com>