drivers: introduce rpmsg, a remote-processor messaging bus
Add a virtio-based IPC bus, which enables kernel users to communicate
with remote processors over shared memory using a simple messaging
protocol.

Assign a local address for every local endpoint that is created,
and bind it to the user's callback. Invoke that callback when the
destination of an inbound message is the user's address.

Signed-off-by: Ohad Ben-Cohen <ohad@wizery.com>