rpmsg_proto: Test flags rather than msg->msg_flags in rpmsg_sock_recvmsg().
From http://linux.die.net/man/2/recvmsg it appears, though is not very clear,
that the flags parameter to sock_recvmsg() is for input, and the msg->msg_flags
field is for output.

"The msg_flags field in the msghdr is set on return of recvmsg()."

The recvmsg function was testing the msg->msg_flags field, which
intermittently tested true for MSG_OOB, causing a spurious return EOPNOTSUPP.

Setting the conditional test to (flags & MSG_OOB) seems to have corrected the
intermittent error condition, as validated by a soak test.

Signed-off-by: G Anthony <a0783926@ti.com>